Move "heavy checks"-related tests to framework_tests.rs
